MP2-1068 E2 section unit costs fixes

See merge request interact/ems/ems!1477

+++ jems-server/src/main/kotlin/io/cloudflight/jems/server/project/entity/partner/budget/unit_cost/ProjectUnitCostRow.kt
@@ -1,13 +1,14 @@
 package io.cloudflight.jems.server.project.entity.partner.budget.unit_cost
 
 import io.cloudflight.jems.server.common.entity.TranslationView
+import java.math.BigDecimal
 
 interface ProjectUnitCostRow: TranslationView {
 
     val id: Long
     val costId: Long
-    val pricePerUnit: Long?
-    val numberOfUnits: Long?
+    val pricePerUnit: BigDecimal?
+    val numberOfUnits: BigDecimal?
     val name: String?
     val description: String?
     val unitType: String?
